prompt_enhancing/pyproject.toml
[project]
name = "archaeo-super-prompt"
version = "0.1.0"
description = "DSPy pipeline to generate several optimized prompt lists for extracting structured information on archaeological legacy raw documents"
authors = [
    {name = "Convolutio",email = "79534557+Convolutio@users.noreply.github.com"}
]
license = {text = "UNLICENSED"}
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "dspy (>=2.6.24,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "bidict (>=0.23.1,<0.24.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "pymupdf (>=1.26.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pandera[mypy] (>=0.24.0,<0.25.0)",
    "feature-engine (>=1.8.3,<2.0.0)",
    "dash (>=3.0.4,<4.0.0)",
    "docling (>=2.39.0,<3.0.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "thefuzz (>=0.22.1,<0.23.0)",
]

[project.scripts]
main = "archaeo_super_prompt.main:main"

[tool.poetry]

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.7"
nbstripout = "^0.8.1"
pyyaml = "^6.0.2"
black = "^25.1.0"
frictionless = "^5.18.1"
nest-asyncio = "^1.6.0"
tqdm-stubs = "^0.2.1"
pandas-stubs = ">=2.2.3.250527,<3.0.0.0"
ruff = "^0.12.3"


[[tool.poetry.source]]
name = "torch-cpu-only"
url = "https://download.pytorch.org/whl"
priority = "supplemental"


[tool.poetry.dependencies]
torch = {source = "torch-cpu-only"}
torchvision = {source="torch-cpu-only"}

[tool.mypy]
disable_error_code = ["import-untyped"]
plugins = ["pandera.mypy"]

[tool.ruff]
line-length=79

[tool.ruff.lint]
extend-select = [
  "UP",  # pyupgrade
  "D",   # pydocstyle
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
